+++ src/Files.Shared/Logger.cs
@@ -7,20 +7,6 @@ using System.Threading.Tasks;
 
 namespace Files.Shared
 {
-    public interface ILogger
-    {
-        void Info(string error, [CallerMemberName] string caller = "");
-        void Info(string info, object obj, [CallerMemberName] string caller = "");
-        void Info(Exception ex, string error = "", [CallerMemberName] string caller = "");
-
-        void Warn(string error, [CallerMemberName] string caller = "");
-        void Warn(Exception ex, string error = "", [CallerMemberName] string caller = "");
-
-        void Error(string error, [CallerMemberName] string caller = "");
-        void Error(Exception ex, string error = "", [CallerMemberName] string caller = "");
-        void UnhandledError(Exception ex, string error = "", [CallerMemberName] string caller = "");
-    }
-
     public class Logger : ILogger
     {
         private readonly ILogWriter writer;
